Update database queries and add new configuration options
This commit updates the database queries in the service file to only fetch certificate requests with a status of 0. It also adds new configuration options for the police check service, ID check service, and address check service. Additionally, the commit introduces a new record type called division_record and modifies the allDivisions function to return an array of division_record objects.